+++ src/Products/MeetingLalouviere/tests/testWFAdaptations.py
@@ -316,6 +316,41 @@ class testWFAdaptations(MeetingLalouviereTestCase, mctwfa):
                 ('item_validation_no_validate_shortcuts', )),
             wa_dependencies)
 
+    def test_pm_ItemDecidedWithReturnToProposingGroup(self):
+        """Override as we have a special behavior, possible to
+           "backTo_itempublished_from_returned_to_proposing_group" when meeting is
+           "published" or "decided"."""
+        if not self._check_wfa_available(
+                ['itemdecided', 'return_to_proposing_group']):
+            return
+        # enable auto itemdecide item when meeting decided
+        cfg = self.meetingConfig
+        actions = list(cfg.getOnMeetingTransitionItemActionToExecute())
+        actions.insert(5,
+                       {'meeting_transition': 'decide',
+                        'item_action': 'itemdecide',
+                        'tal_expression': ''})
+        cfg.setOnMeetingTransitionItemActionToExecute(actions)
+        self._activate_wfas(('itemdecided', 'return_to_proposing_group'))
+        self.changeUser('pmManager')
+        item = self.create('MeetingItem', decision=self.decisionText)
+        meeting = self.create('Meeting')
+        self.presentItem(item)
+        self.do(item, 'return_to_proposing_group')
+        self.assertEqual(self.transitions(item),
+                         ['backTo_presented_from_returned_to_proposing_group'])
+        self.freezeMeeting(meeting)
+        self.assertEqual(self.transitions(item),
+                         ['backTo_itemfrozen_from_returned_to_proposing_group'])
+        self.decideMeeting(meeting)
+        # XXX begin change by MeetingLalouviere
+        self.assertEqual(self.transitions(item),
+                         ['backTo_itemdecided_from_returned_to_proposing_group',
+                          'backTo_itempublished_from_returned_to_proposing_group'])
+        # XXX end change by MeetingLalouviere
+        self.do(item, 'backTo_itemdecided_from_returned_to_proposing_group')
+        self.assertEqual(item.query_state(), 'itemdecided')
+
 
 def test_suite():
     from unittest import TestSuite, makeSuite